Avoid missing switch changes during reboot
To ensure that changes to the boot steering
switch setting are not missed during reboot the
expected value of the switch for the image running
is OR’d in to the value read from the port.

If the switch position changed during the reboot
and interrupt should be raised immediately.