Register tests and set minimum testlevel
All test scripts have been updated to register test cases in the
constructor of the Tester, and to set the the minimum testlevel of the
CombinedTester object. This will enable features of the xmostest
framework to work correctly when skipping tests due the current
test run level, and when generating testplans or getting total test
case counts.

The dictionary of test cases has been improved and simplified to make
it simpler to add/remove test cases and to control when different test
cases run.

Sample frequency has been replaced with ‘sample rate’ to avoid any
confusion in the Testers.